requirements.txt
# ~/life requirements
# Install to venv: ~/life/.venv/bin/pip install -r requirements.txt

# Google APIs (OAuth2 + Calendar, Gmail, Drive, Sheets)
google-auth-oauthlib
google-api-python-client
google-auth-httplib2

# Environment / config
python-dotenv

# HTTP
requests
